requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.0.3
Flask-Cors==4.0.1
pillow==10.3.0
timm==1.0.3
torch==2.3.0
torchvision==0.18.0
